Avoid disabling 4267 and 4244 warnings
4267 - 'var' : conversion from 'size_t' to 'type', possible loss of data.
4244 - 'argument' : conversion from 'type1' to 'type2', possible loss of data.

+++ src/tbbmalloc/Customize.h
@@ -47,8 +47,9 @@
 #define MALLOC_ITT_RELEASE_RESOURCES()  ((void)0)
 #endif
 
-inline intptr_t BitScanRev(uintptr_t x) {
-    return x == 0 ? -1 : static_cast<intptr_t>(tbb::detail::log2(x));
+// TODO: Consider conforming to _BitScanReverse interface, an MSVC intrinsic
+inline int BitScanRev(uintptr_t x) {
+    return x == 0 ? -1 : static_cast<int>(tbb::detail::log2(x));
 }
 
 template<typename T>
